+++ app/routes/course_material.py
@@ -3,7 +3,7 @@ from typing import List
 from fastapi import APIRouter, HTTPException, status, Query, Body
 
 from app.models.course_material import CourseMaterial
-from app.utils.azure_blob_uploader import AzureBlobUploader
+from app.utils.azure_blob_service import AzureBlobUploader
 
 router = APIRouter()
